MAGECLOUD-1203: Different logging systems support
